Refactor modules back to simple datasets module
