--- espei/datasets/__init__.py
@@ -1,2 +1,0 @@
-from .dataset_models import *
-from .db import *
